requirements.txt
Django==1.10.6
django-environ==0.4.3
django-taggit==0.22.0
django-taggit-serializer==0.1.5
djangorestframework==3.6.2
factory-boy==2.9.2
Faker==0.7.12
httpie==0.9.9
inflection==0.3.1
psycopg2==2.7.1
py==1.4.34
Pygments==2.2.0
pytest==3.2.1
pytest-django==3.1.2
pytest-factoryboy==1.3.1
python-dateutil==2.6.0
requests==2.14.2
six==1.10.0
